add e2e tests
