+++ packages/plugin-e2e/tests/as-admin-user/openfeature/openFeature-compatibility.spec.ts
@@ -1,0 +1,8 @@
+import { expect, test } from '../../../src';
+
+// No featureToggles specified - should work unchanged
+test('should work without featureToggles option', async ({ page }) => {
+  // Simply verify the page loads without errors
+  await page.goto('/');
+  await expect(page).toHaveURL(/.*\//);
+});
